+++ src/main.h
@@ -21,14 +21,16 @@
 #define _MAIN_H
 
 #include <fcntl.h>
+#include "config.h"
+#ifndef FEATURE_NO_XEMBED
 #include <gtk/gtkx.h>
+#endif
 #include <stdio.h>
 #include <webkit2/webkit2.h>
 #include "shortcut.h"
 #include "handler.h"
 #include "file-storage.h"
 
-#include "config.h"
 
 #define LENGTH(x) (sizeof x / sizeof x[0])
 #define OVERWRITE_STRING(t, s) {if (t) g_free(t); t = g_strdup(s);}
@@ -246,6 +248,7 @@ struct Client {
          * client base. */
         GHashTable              *settings;
         guint                   scrollstep;
+        guint                   scrollmultiplier;
         gboolean                input_autohide;
         gboolean                incsearch;
         gboolean                prevent_newwindow;
@@ -271,12 +274,15 @@ struct Client {
 struct Vimb {
     char        *argv0;
     Client      *clients;
+#ifndef FEATURE_NO_XEMBED
     Window      embed;
+#endif
     GHashTable  *modes;             /* all available browser main modes */
     char        *configfile;        /* config file given as option on startup */
     char        *files[FILES_LAST];
     FileStorage *storage[STORAGE_LAST];
     char        *profile;           /* profile name */
+    GSList      *cmdargs;           /* ex commands given asl --command, -C option */
     struct {
         guint   history_max;
         guint   closed_max;
@@ -284,6 +290,8 @@ struct Vimb {
     GtkCssProvider *style_provider;
     gboolean    no_maximize;
     gboolean    incognito;
+
+    WebKitWebContext *webcontext;
 };
 
 gboolean vb_download_set_destination(Client *c, WebKitDownload *download,