+++ src/util.c
@@ -48,13 +48,13 @@ static gboolean match_list(const char *pattern, int patlen, const char *subject)
  *
  * Returned path must be freed.
  */
-char *util_build_path(State state, const char *path, const char *dir)
+char *util_build_path(const char *path, const char *dir)
 {
     char *fullPath = NULL, *fexp, *dexp, *p;
     int expflags   = UTIL_EXP_TILDE|UTIL_EXP_DOLLAR;
 
     /* if the path could be expanded */
-    if ((fexp = util_expand(state, path, expflags))) {
+    if ((fexp = util_expand(path, expflags))) {
         if (*fexp == '/') {
             /* path is already absolute, no need to use given dir - there is
              * no need to free fexp, because this should be done by the caller
@@ -62,7 +62,7 @@ char *util_build_path(State state, const char *path, const char *dir)
             fullPath = fexp;
         } else if (dir && *dir) {
             /* try to expand also the dir given - this may be ~/path */
-            if ((dexp = util_expand(state, dir, expflags))) {
+            if ((dexp = util_expand(dir, expflags))) {
                 /* use expanded dir and append expanded path */
                 fullPath = g_build_filename(dexp, fexp, NULL);
                 g_free(dexp);
@@ -161,7 +161,7 @@ gboolean util_create_tmp_file(const char *content, char **file)
  *
  * Returned path must be g_freed.
  */
-char *util_expand(State state, const char *src, int expflags)
+char *util_expand(const char *src, int expflags)
 {
     const char **input = &src;
     char *result;
@@ -169,7 +169,7 @@ char *util_expand(State state, const char *src, int expflags)
     int flags    = expflags;
 
     while (**input) {
-        util_parse_expansion(state, input, dst, flags, "\\");
+        util_parse_expansion(input, dst, flags, "\\");
         if (VB_IS_SEPARATOR(**input)) {
             /* after space the tilde expansion is allowed */
             flags = expflags;
@@ -335,6 +335,30 @@ char *util_file_pop_line(const char *file, int *item_count)
 char *util_get_config_dir(void)
 {
     char *path = g_build_filename(g_get_user_config_dir(), PROJECT, vb.profile, NULL);
+    create_dir_if_not_exists(path);
+
+    return path;
+}
+
+/**
+ * Retrieves the data directory path according to current used profile.
+ * Returned string must be freed.
+ */
+char *util_get_data_dir(void)
+{
+    char *path = g_build_filename(g_get_user_data_dir(), PROJECT, vb.profile, NULL);
+    create_dir_if_not_exists(path);
+
+    return path;
+}
+
+/**
+ * Retrieves the cache directory path according to current used profile.
+ * Returned string must be freed.
+ */
+char *util_get_cache_dir(void)
+{
+    char *path = g_build_filename(g_get_user_cache_dir(), PROJECT, vb.profile, NULL);
     create_dir_if_not_exists(path);
 
     return path;
@@ -556,7 +580,7 @@ gboolean util_fill_completion(GtkListStore *store, const char *input, GList *src
  * Fills file path completion entries into given list store for also given
  * input.
  */
-gboolean util_filename_fill_completion(State state, GtkListStore *store, const char *input)
+gboolean util_filename_fill_completion(GtkListStore *store, const char *input)
 {
     gboolean found = FALSE;
     GError *error  = NULL;
@@ -568,9 +592,8 @@ gboolean util_filename_fill_completion(State state, GtkListStore *store, const c
     input_basename = last_slash ? last_slash + 1 : input;
     input_dirname  = g_strndup(input, input_basename - input);
     real_dirname   = util_expand(
-        state,
         *input_dirname ? input_dirname : ".",
-        UTIL_EXP_TILDE|UTIL_EXP_DOLLAR|UTIL_EXP_SPECIAL
+        UTIL_EXP_TILDE|UTIL_EXP_DOLLAR
     );
 
     dir = g_dir_open(real_dirname, 0, &error);
@@ -651,7 +674,7 @@ char *util_js_result_as_string(WebKitJavascriptResult *result)
  * @quoteable: String of chars that are additionally escapable by \.
  * Returns TRUE if input started with expandable pattern.
  */
-gboolean util_parse_expansion(State state, const char **input, GString *str,
+gboolean util_parse_expansion(const char **input, GString *str,
         int flags, const char *quoteable)
 {
     GString *name;
@@ -725,12 +748,6 @@ gboolean util_parse_expansion(State state, const char **input, GString *str,
         /* variable are expanded even if they do not exists */
         expanded = TRUE;
         g_string_free(name, TRUE);
-    } else if (flags & UTIL_EXP_SPECIAL && **input == '%') {
-        if (state.uri) {
-            /* TODO check for modifiers like :h:t:r:e */
-            g_string_append(str, state.uri);
-            expanded = TRUE;
-        }
     }
 
     if (!expanded) {
@@ -748,7 +765,6 @@ gboolean util_parse_expansion(State state, const char **input, GString *str,
             } else if (strchr(quoteable, **input)
                 || (flags & UTIL_EXP_TILDE && **input == '~')
                 || (flags & UTIL_EXP_DOLLAR && **input == '$')
-                || (flags & UTIL_EXP_SPECIAL && **input == '%')
             ) {
                 /* escaped char becomes only char */
                 g_string_append_c(str, **input);
@@ -1106,3 +1122,39 @@ static gboolean match_list(const char *pattern, int patlen, const char *subject)
     }
 }
 
+/**
+ * Get the time span to given string like '1y5dh' (one year and five days and
+ * one hour).
+ */
+GTimeSpan util_string_to_timespan(const char *input)
+{
+    unsigned int multiplier = 0;
+    GTimeSpan sum           = 0;
+    gboolean hasmultiplier  = FALSE;
+
+    while (*input) {
+        if (VB_IS_DIGIT(*input)) {
+            multiplier    = multiplier * 10 + (*input - '0');
+            /* Use separate variable to differentiate between no multiplier
+             * set (is same as 1) or ultiplier of '0'. */
+            hasmultiplier = TRUE;
+        } else {
+            GTimeSpan s = 0;
+            switch (*input) {
+                case 'y': s = 365 * G_TIME_SPAN_DAY; break;
+                case 'w': s = 7 * G_TIME_SPAN_DAY; break;
+                case 'd': s = G_TIME_SPAN_DAY; break;
+                case 'h': s = G_TIME_SPAN_HOUR; break;
+                case 'm': s = G_TIME_SPAN_MINUTE; break;
+                case 's': s = G_TIME_SPAN_SECOND; break;
+            }
+            sum += s * (hasmultiplier ? multiplier : 1);
+            /* Unset multiplier for th epossible next multiplier in input */
+            multiplier    = 0;
+            hasmultiplier = FALSE;
+        }
+        input++;
+    }
+
+    return sum;
+}